requirements.txt
# Core dependencies
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.7.0
whitenoise==6.9.0

# NLP & Summarization
nltk==3.9.1
sumy==0.11.0
textblob==0.19.0
regex==2024.11.6

# API & Requests
requests==2.32.3
urllib3==2.2.3

# Database (Only if using MySQL)
mysqlclient==2.2.6
PyMySQL==1.1.1

# Optional (Keep only if needed)
gunicorn==23.0.0   # For running Django on Render
selenium==4.27.1   # Keep only if using Selenium for web scraping
